Feat : Proof-of-Custody (chip-backed authenticité) + durcissement sécurité
Nouveau protocole Proof-of-Custody, réutilisant le pipeline Attestcoin déjà
audité pour prouver l'absence de substitution d'un actif physique chippé
(EIP-5791), sans oracle ni jury pour le cas courant :
- CustodyRouter.sol (Sepolia) : attestPossession (Origin/Delivery, signature
  de puce vérifiée par ECDSA.recover).
- SellerPassport.sol (Creditcoin) : casier de faits côté vendeur, miroir de
  CreditPassport ; waivesBond() dispense de caution après 4 livraisons
  propres.
- PayGoEscrow.sol : settleCustody (mêmes 5 vérifications que settle),
  caution optionnelle en CTC natif (jamais indexée sur price — pas
  d'oracle CTC/payToken), pattern pull pour claimAsset/withdrawAsset
  (liste blanche d'actifs) au lieu du push automatique.

Audit multi-agents (skill sc-audit, 4 hunters + 8 refuters) sur cette
même fonctionnalité : 3 vulnérabilités confirmées par PoC, corrigées :
- CreditPassport.volume (uint128) débordait et bloquait définitivement
  settle() pour un acheteur ciblé ; payToken n'était jamais mis en liste
  blanche. Fix : volume en uint256, allowedPayTokens ajouté.
- withdrawBond n'avait aucun chemin de récupération si le destinataire ne
  pouvait pas recevoir un transfert brut. Fix : scindé en withdrawBond
  (résolution, jamais d'appel externe) + claimBond (pull par le
  destinataire).
- _applyCustodyLog ne vérifiait jamais le soumetteur réel (msg.sender) des
  attestations. Fix : submitter == seller/buyer selon le rôle — corrige le
  squattage par un tiers ; un résiduel de conception (auto-forgerie par un
  vendeur contrôlant deux adresses) reste documenté, non corrigeable par un
  patch seul (docs/AUDIT.md, docs/08-proof-of-custody.md).

47/47 tests passent (25 -> 47). Détail complet : audit/00-recon.md,
audit/findings/, docs/AUDIT.md.

Redéployé sur Sepolia + Creditcoin CC3 testnet (v3) ; adresses à jour dans
README.md et .env.

+++ contracts/CreditPassport.sol
@@ -6,7 +6,11 @@ pragma solidity ^0.8.23;
 ///         written by PayGoEscrow, and read back by PayGoEscrow itself to size the next deposit.
 ///         One token per address, tokenId = uint160(owner), minted on first fact, never transferable.
 contract CreditPassport {
-    struct Record { uint32 honored; uint32 defaulted; uint128 volume; }
+    /// @dev `volume` is `uint256`, not a narrower type: it's purely informational (never gates
+    ///      `depositBps`), and a narrower type invites exactly the truncating-cast-then-overflow DoS a
+    ///      security pass found here (SC-AUDIT-01) — `amount` is attacker-controlled ERC20 input with no
+    ///      protocol-level ceiling, so any fixed-width accumulator is a griefing vector, not just uint128.
+    struct Record { uint32 honored; uint32 defaulted; uint256 volume; }
 
     address public immutable escrow;
     mapping(address => Record) public records;
@@ -21,7 +25,7 @@ contract CreditPassport {
         require(msg.sender == escrow, "escrow only");
         Record storage r = records[buyer];
         if (r.honored + r.defaulted == 0) { emit Transfer(address(0), buyer, uint160(buyer)); emit Locked(uint160(buyer)); }
-        if (ok) { r.honored++; r.volume += uint128(amount); } else r.defaulted++;
+        if (ok) { r.honored++; r.volume += amount; } else r.defaulted++;
         emit Fact(buyer, ok, amount);
     }
